branch for registering crash to be implemented
This branch was created to separate the crash implementation.
It also has two separate Jason files for the implementation
of the pains and the sensors.
they are there because I wanted to implement ultra sensors
instead of infrared it was not successful yet.

PS I turned her car into a tank ";>"