finish implementing issue #61
The infrared has now been switch to ultra addUltSensors.
The car is also been upgraded so now in the stop zones
when obstacles ahead of it or behind it will not allow the car to go
forward or backwards.
And finally because we loop through the sensors we no longer
have an issue of sensor overload.

Issue #61 is now closed